Update testStore; Change to folder with models
(Update references)
