+++ src/components/layouts/EvaluationLayout.svelte
@@ -6,7 +6,7 @@
   import { getContext, onDestroy, onMount, setContext } from 'svelte';
   import { outcomeValueStore } from '@app/stores/earl/resultStore/index.js';
   import subjects, { TestSubjectTypes } from '@app/stores/earl/subjectStore/index.js';
-  import testStore from '@app/stores/earl/testStore.js';
+  import testStore from '@app/stores/earl/testStore/index.js';
 
   const { scopeStore } = getContext('app');
   // Initialize
