+++ src/components/pages/Evaluation/SummaryPage.svelte
@@ -56,7 +56,7 @@
 
 <script>
   import { getContext } from 'svelte';
-  import tests from '@app/stores/earl/testStore.js';
+  import tests from '@app/stores/earl/testStore/index.js';
 
   import Details from '@app/components/Details.svelte';
   import Page from '@app/components/Page.svelte';
